Move LazyLoadingGallery to viewer/gallert/component
